refactor: remove unnecessary `base` package from entity imports
- Updated all entity imports from `entity.base` to `entity` across the project.
- Added user-friendly API methods in `ChannelService` for console operations.
