+++ discordit/src/main/java/com/sprint/mission/discodeit/repository/file/FileChannelRepository.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.repository.file;
 
-import com.sprint.mission.discodeit.entity.base.Channel;
-import com.sprint.mission.discodeit.entity.base.User;
+import com.sprint.mission.discodeit.entity.Channel;
+import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.enums.ChannelScope;
 import com.sprint.mission.discodeit.exceptions.ChannelAlreadyExistsException;
 import com.sprint.mission.discodeit.exceptions.ChannelNotFoundException;
