+++ discordit/src/main/java/com/sprint/mission/discodeit/service/ChannelService.java
@@ -5,7 +5,8 @@ import com.sprint.mission.discodeit.dto.channel.request.PrivateChannelCreateRequ
 import com.sprint.mission.discodeit.dto.channel.request.PublicChannelCreateRequestDto;
 import com.sprint.mission.discodeit.dto.channel.response.ChannelResponseDto;
 import com.sprint.mission.discodeit.dto.user.response.UserResponseDto;
-import com.sprint.mission.discodeit.entity.base.User;
+import com.sprint.mission.discodeit.enums.ChannelType;
+import com.sprint.mission.discodeit.entity.User;
 
 import java.util.List;
 import java.util.UUID;
@@ -32,4 +33,16 @@ public interface ChannelService {
     ChannelResponseDto getNth(int index);
 
     List<ChannelResponseDto> getAllByUserId(String userId);
+
+    // 콘솔 편의용 API (식별/조회 간소화)
+    List<UUID> getAllIds();
+    List<UUID> getAllIdsByUserId(String userId);
+    String getDisplayName(UUID uuid);
+    ChannelType getType(UUID uuid);
+
+    // 콘솔 편의용 멤버 수정 (userId 기반)
+    void addMember(UUID uuid, String userId);
+    void addModerator(UUID uuid, String userId);
+    void deleteMember(UUID uuid, String userId);
+    void deleteModerator(UUID uuid, String userId);
 }
